requirements.txt
fastapi
uvicorn
qdrant-client
Pillow
python-multipart
python-dotenv
openai
requests
scikit-learn
numpy

